// reelease-ai-next/src/types/ecommerceCatalogue.ts
export interface CatalogueModelItem {
  id: string
  name: string
  description: string
  image_url: string
  isFromLibrary?: boolean
}

export interface CatalogueProductItem {
  id: string
  name: string
  image_url: string
}

export type CataloguePickerMode = 'otherCharacter' | 'product' | null